Refactor: Improve vector storage and management
This commit introduces significant refactoring to the vector storage and management system. Key changes include:

*   **Introduction of `VectorManagementService`**: A new service layer is added to abstract vector operations, providing a unified interface for interacting with different vector databases.
*   **Enhanced `InMemoryVectorDatabaseService`**: The in-memory vector database has been significantly improved with new functionalities such as `updateVector`, `getVector`, `getVectorByEntity`, `searchByEntityType`, `removeVectorById`, `batchStoreVectors`, `batchUpdateVectors`, `batchRemoveVectors`, `getVectorsByEntityType`, `getVectorCountByEntityType`, `vectorExists`, `clearVectors`, and `clearVectorsByEntityType`. The internal storage has also been optimized using `VectorRecord` for better data representation.
*   **Updated `LuceneVectorDatabaseService`**: The Lucene-based vector database service has been updated to align with the new `VectorRecord` structure and to use `long` for delete counts, improving consistency.
*   **Placeholder for `PineconeVectorDatabaseService`**: The Pinecone integration is still a placeholder but has been updated to reflect the new `VectorRecord` structure and to provide more informative warnings about its unimplemented status.
*   **New Test Case `VectorStorageMigrationTest`**: A new test class is introduced to specifically verify the vector storage migration and the functionality of the `VectorManagementService`, including batch operations and existence checks.

These changes aim to provide a more robust, flexible, and scalable vector management system, paving the way for future integrations and improvements.

+++ ai-infrastructure-module/ai-infrastructure-core/src/main/java/com/ai/infrastructure/rag/LuceneVectorDatabaseService.java
@@ -248,7 +248,7 @@ public class LuceneVectorDatabaseService implements VectorDatabaseService {
             
             // Delete by entityId and entityType
             Term term = new Term("entityId", entityId);
-            int deletedCount = indexWriter.deleteDocuments(term);
+            long deletedCount = indexWriter.deleteDocuments(term);
             indexWriter.commit();
             
             // Refresh reader
@@ -323,7 +323,7 @@ public class LuceneVectorDatabaseService implements VectorDatabaseService {
             
             // First remove the existing vector
             Term term = new Term("vectorId", vectorId);
-            int deletedCount = indexWriter.deleteDocuments(term);
+            long deletedCount = indexWriter.deleteDocuments(term);
             
             if (deletedCount > 0) {
                 // Store the updated vector
@@ -404,7 +404,7 @@ public class LuceneVectorDatabaseService implements VectorDatabaseService {
             log.debug("Removing vector {} from Lucene", vectorId);
             
             Term term = new Term("vectorId", vectorId);
-            int deletedCount = indexWriter.deleteDocuments(term);
+            long deletedCount = indexWriter.deleteDocuments(term);
             indexWriter.commit();
             
             // Refresh reader
